converting floor to int

+++ DensityPlotting.py
@@ -1323,7 +1323,7 @@ def AnalyzeBinary(beginStep, lastStep, dmFiles, gasFiles, savingDir, outputDir, 
     Qzz = multiprocessing.Array('f', [0.0 for i in workingRange])
 
     #chunkSize = (lastStep - beginStep) / 8
-    chunkSize = math.floor(len(workingRange) / cpus)
+    chunkSize = int(math.floor(len(workingRange) / cpus))
 
     if chunkSize == 0:
         if lastStep - beginStep == 0:
